requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1695154477682/work
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==2.2.5
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1637590993074/work
gunicorn==21.2.0
h5py @ file:///D:/bld/h5py_1617739645326/work
idna==3.4
importlib-metadata @ file:///D:/bld/importlib-metadata_1653252990345/work
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
libclang==16.0.6
Mako @ file:///home/conda/feedstock_root/build_artifacts/mako_1668568582731/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1696627721496/work
MarkupSafe==2.1.3
numpy @ file:///D:/bld/numpy_1649806536290/work
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pickle-mixin==1.0.2
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygpu==0.7.6
pyreadline @ file:///D:/bld/pyreadline_1655703530615/work
PyYAML @ file:///D:/bld/pyyaml_1648757357341/work
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy @ file:///C:/bld/scipy_1637806996411/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1682317048417/work
Theano @ file:///D:/bld/theano_1647410736059/work
threadpoolctl==3.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3==2.0.6
Werkzeug==2.2.3
wrapt @ file:///D:/bld/wrapt_1602463555786/work
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
